+++ components/Ledger/Login.tsx
@@ -85,6 +85,31 @@ const Login = (): JSX.Element => {
         setCurrentUser(getLocalUser);
     });
 
+    const checkIfAccount = async (address: string) => {
+        const accountAddress = { address: address };
+        try {
+            const isAccount = await Ledger.isAccount(accountAddress);
+            if (isAccount === true) {
+                return true;
+            } else {
+                setLoading(true);
+                setErrorMessage(
+                    `Can't find an account with address ${address} on the chain. Please accept the connection and sign the transaction on your Ledger deivce to create an account. `
+                );
+                const createAccount = await Ledger.createAccount(accountAddress);
+                if (createAccount === true) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+        } catch (e) {
+            setLoading(true);
+            setErrorMessage(Ledger.checkLedgerErrors(e.message));
+            setRetry(true);
+        }
+    };
+
     const handleLogin = async () => {
         if (!currentUser || currentUser === '' || currentUser === 'undefined') {
             setLoading(true);
@@ -106,9 +131,11 @@ const Login = (): JSX.Element => {
                     setErrorMessage('Please accept the connection in your Ledger device. ');
                     try {
                         const userAddress = await Ledger.getAddress();
-                        localStorage.setItem('currentUserAddress', userAddress);
-                        setCurrentUser(userAddress);
-                        setOpen(false);
+                        if (await checkIfAccount(userAddress)) {
+                            localStorage.setItem('currentUserAddress', userAddress);
+                            setCurrentUser(userAddress);
+                            setOpen(false);
+                        }
                     } catch (e) {
                         setLoading(true);
                         setErrorMessage(Ledger.checkLedgerErrors(e.message));