+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 * Added xxsmall and xsmall row size option in `AccountList`
 * Handled NAN values showing in some Validator Group elements 
 * Updated page value in `DepositList`
+* Updated Ledger Login
 
 ## v0.1.3